Split the schema management into dedicated files, adjusted appendEvent method and stream structure to handle onconclict
